issue #22: add method from _package to PackageDependenciesInstaller class
