+++ src/kvo/devtools/setuppackage.py
@@ -1,5 +1,5 @@
 from .package import Package
-from .dependencies import install_dependencies
+from .packagedependencies import PackageDependenciesInstaller
 from .packageindexes import PackageIndex
 
 
@@ -9,4 +9,5 @@ async def setup_package(package: Package, package_index: PackageIndex | None = N
     This method is an asynchronous wrapper for the download and install_deps methods.
     """
     await package.download()
-    await install_dependencies(package, package_index)
+    installer = PackageDependenciesInstaller.from_package(package, package_index)
+    await installer.install_dependencies()
